Updating the network commands using local action execution, updating scoping, and removing network data mixin.

+++ app/data/network/management/commands/_network/rm.py
@@ -3,7 +3,6 @@ from systems.command import types, mixins
 
 class RemoveCommand(
     mixins.op.RemoveMixin,
-    mixins.data.NetworkMixin,
     types.NetworkActionCommand
 ):
     def get_description(self, overview):
@@ -36,5 +35,12 @@ velit. Aenean sit amet consequat mauris.
         self.confirmation()       
 
     def exec(self):
-        self.network.provider.destroy_network()
-        self.exec_rm(self._network, self.network.name)
+        if self.check_exists(self._network, self.network.name):
+            self.exec_local('subnet clear', {
+                'network_name': self.network.name
+            })
+            self.exec_local('firewall clear', {
+                'network_name': self.network.name
+            })
+            self.network.provider.destroy_network()
+            self.exec_rm(self._network, self.network.name)
